veachron/requirements.txt
Werkzeug <= 2.1.2   
flask == 2.1.2   
flask-restx >= 0.5.1 
flask-markdown == 0.3
psycopg2-binary >= 2.9.4